+++ migrations_test.go
@@ -30,7 +30,7 @@ func TestMigrations(t *testing.T) {
 }
 
 func TestLatestSchema(t *testing.T) {
-	assert.Equal(t, 19, latestSchema())
+	assert.Equal(t, 20, latestSchema())
 }
 
 func TestNoDuplicateMigrations(t *testing.T) {