+++ pkg/heroku/app_test.go
@@ -280,7 +280,7 @@ func TestAppUpdateSuccess(t *testing.T) {
 	ts, handler, c := newTestServerAndClient(t, appUpdateRequest)
 	defer ts.Close()
 
-	app, err := c.AppUpdate("example", &appUpdateRequestOptions)
+	app, err := c.AppUpdate("example", &appUpdateRequestOptions, "message")
 	if err != nil {
 		t.Fatal(err)
 	}